Moved config.py -> src/hydrarpr
