+++ src/hydrarpr/render_studio/config.py
@@ -15,8 +15,6 @@
 from pathlib import Path
 import uuid
 
-logging_level = 'DEBUG'     # available levels: 'DEBUG', 'INFO', 'WARN', 'ERROR', 'CRITICAL'
-logging_backups = 5
 render_studio_dir = Path.home() / "Documents/AMD RenderStudio Home"
 user_id = f"BlenderUser_{uuid.uuid4()}"
 server_url = ""
